% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cfu.R
\name{calculate_cfu}
\alias{calculate_cfu}
\title{Calculate the cfus based on the dilution factor and resuspended volume}
\usage{
calculate_cfu(
  df,
  dilution_factor,
  resuspend_volume_ml,
  percent = 1,
  dilution_column = "dilution",
  CFU_column = "CFUs"
)
}
\arguments{
\item{df}{dataframe that contains information on the cfu counts and corresponding dilutions}

\item{dilution_factor}{the dilution factor used to perform dilutions (e.g., 5 fold dilutions)}

\item{resuspend_volume_ml}{the volume that the organ is resuspended in mL (e.g., 0.5 mL)}

\item{percent}{percentage of the organ used to plate as a fraction/decimal, default is 1 (whole organ)}

\item{dilution_column}{name of the dilution column in the input dataframe}

\item{CFU_column}{name of the CFU column in the input dataframe}
}
\value{
df
}
\description{
Calculate the cfus based on the dilution factor and resuspended volume
}
\examples{
calculate_cfu(CFU_one_dilution, 5, 0.5, .5, "dilution", "CFUs")


}
