% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CFU_excel_raw}
\alias{CFU_excel_raw}
\title{Raw CFU data}
\format{
A data frame with 46 rows and 7 variables:
\describe{
  \item{group}{group number to which each mouse belongs}
  \item{mouse}{mouse name}
  \item{organ}{organ name of each sample}
  \item{dilution_0}{number of CFUs present for the undiluted sample}
  \item{dilution_1}{number of CFUs present for dilution 1}
  \item{dilution_2}{number of CFUs present for dilution 2}
  \item{dilution_3}{number of CFUs present for dilution 3}
  ...
}
}
\usage{
CFU_excel_raw
}
\description{
A dataset containing the Colony Forming Units (CFUs) for different mice
at various dilutions
}
\keyword{datasets}
