% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_one_dilution.R
\name{pick_one_dilution}
\alias{pick_one_dilution}
\title{Pick one dilution to perform CFU calculations on}
\usage{
pick_one_dilution(df, CFU_column, grouping_columns)
}
\arguments{
\item{df}{dataframe that contains columns: dilution, CFU, and grouping variables}

\item{CFU_column}{the name of the CFU column in the dataframe}

\item{grouping_columns}{the columns names to group the data based on (i.e., Group, Timepoint)}
}
\value{
df
}
\description{
Pick one dilution to perform CFU calculations on
}
\examples{
pick_one_dilution(CFU_raw_formatted, "CFUs", c("group", "organ", "replicate"))

}
