% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CFU_raw_formatted}
\alias{CFU_raw_formatted}
\title{CFU data formatted after using tidy_cfu}
\format{
A data frame with 46 rows and 7 variables:
\describe{
  \item{group}{group number to which each mouse belongs}
  \item{mouse}{mouse name}
  \item{organ}{organ name of each sample}
  \item{dilution}{the dilution at which each CFU count is counted}
  \item{CFUs}{number of CFUs present for each of the corresponding dilutions}
  ...
}
}
\usage{
CFU_raw_formatted
}
\description{
A dataset containing the Colony Forming Units (CFUs) for different mice
at various dilutions
}
\keyword{datasets}
