% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_cfu.R
\name{tidy_CFU}
\alias{tidy_CFU}
\title{Clean CFU data and selects all columns that start with dilution}
\usage{
tidy_CFU(df)
}
\arguments{
\item{df}{dataframe that contains all of the raw CFU data with the dilutions}
}
\value{
df
}
\description{
Clean CFU data and selects all columns that start with dilution
}
\examples{
tidy_CFU(CFU_data)

}
